server/requirements.txt
Django
psycopg2-binary
djangorestframework
pandas
nltk
numpy
Pillow
requests
django-cors-headers
matplotlib
scikit-learn
openai
tqdm
